+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated BOS server setup file from Python 3.6 to Python 3.11
 - Updated from BOA 1.3 to 1.4
+- For POST requests to `/v1/session` that include a `templateUuid` field, convert that field to a `templateName` field (if
+  that field is not also specified) and delete the `templateUuid` field, before creating  a `V1Session` object.
 ### Fixed
 - Fixed a window during power-on operations which could lead to an incorrect status in larger systems
 - Updated API spec so that it accurately describes the actual implementation: